+++ routes/reportes.router.js
@@ -10,6 +10,7 @@ const router = express.Router();
 router.get("/",  reportes.main);
 router.get("/download" , reportes.descargar);
 router.get("/PartiCurso/:curso" , reportes.ParticipantesEnCursos);
+router.get("/directorioEmpresas" , reportes.directorioEmpresas);